Made a data folder, added two files in it, Edited the Projects jsx accordingly
