+++ src/App.jsx
@@ -3,7 +3,7 @@ import Planofaction from './Components/Planofaction.jsx'
 import Footer from './Components/Footer.jsx'
 import Events from './Components/Events.jsx'
 import Navbar from './Components/Navbar.jsx'
-import Projects from './pages/Projects';
+import Projects from './Components/Projects.jsx'
 
 function App() {
   return (
