+++ src/Components/Events.jsx
@@ -69,7 +69,7 @@ const posts = [
 ];
 export default function Events() {
   return (
-    <div className="bg-[#000000] text-white py-24 sm:py-32">
+    <div className="bg-gray-900 text-white py-24 sm:py-32">
       <div className="mx-auto max-w-7xl px-6 lg:px-8">
         <div className="mx-auto max-w-2xl text-center">
           <h2 className="text-3xl font-bold tracking-tight">
