+++ src/Components/Projects.jsx
@@ -1,0 +1,118 @@
+import React from 'react';
+import LiveProjectsData from '../Data/LiveProjectsData'
+import PassionProjectsData from '../Data/PassionPeojectsData'
+
+export default function Projects() {
+  return (
+    <div className="bg-white py-24 sm:py-32">
+      <div className="max-w-7xl px-6 lg:px-8">
+        <div className="max-w-2xl text-left mx-auto mb-8">
+          <h2 className="text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
+            From the blog
+          </h2>
+          <p className="mt-2 text-lg leading-8 text-gray-600">
+            Learn how to grow your business with our expert advice.
+          </p>
+        </div>
+        <div className="mt-16 grid max-w-2xl auto-rows-fr grid-cols-1 gap-8 sm:mt-20 lg:mx-0 lg:max-w-none lg:grid-cols-3">
+          {LiveProjectsData.map((post) => (
+            <article
+              key={post.id}
+              className="relative flex flex-col justify-end overflow-hidden rounded-2xl bg-gray-900 px-8 pb-8 pt-80 sm:pt-48 lg:pt-80"
+            >
+              <img
+                src={post.imageUrl}
+                alt=""
+                className="absolute inset-0 z-0 h-full w-full object-cover"
+              />
+              <div className="absolute inset-0 z-10 bg-gradient-to-t from-gray-900 via-gray-900/40" />
+              <div className="absolute inset-0 z-10 rounded-2xl ring-1 ring-inset ring-gray-900/10" />
+
+              <div className="flex flex-wrap items-center gap-y-1 overflow-hidden text-sm leading-6 text-gray-300">
+                <time dateTime={post.datetime} className="mr-8">
+                  {post.date}
+                </time>
+                <div className="-ml-4 flex items-center gap-x-4">
+                  <svg
+                    viewBox="0 0 2 2"
+                    className="-ml-0.5 h-0.5 w-0.5 flex-none fill-white/50"
+                  >
+                    <circle cx={1} cy={1} r={1} />
+                  </svg>
+                  <div className="flex gap-x-2.5">
+                    <img
+                      src={post.author.imageUrl}
+                      alt=""
+                      className="h-6 w-6 flex-none rounded-full bg-white/10"
+                    />
+                    {post.author.name}
+                  </div>
+                </div>
+              </div>
+              <h3 className="mt-3 text-lg font-semibold leading-6 text-white">
+                <a href={post.href}>
+                  <span className="absolute inset-0" />
+                  {post.title}
+                </a>
+              </h3>
+            </article>
+          ))}
+        </div>
+      </div>
+      <div className="max-w-7xl px-6 lg:px-8">
+        <div className="max-w-2xl text-left mx-auto">
+          <h2 className="text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
+            From the blog
+          </h2>
+          <p className="mt-2 text-lg leading-8 text-gray-600">
+            Learn how to grow your business with our expert advice.
+          </p>
+        </div>
+        <div className="mt-16 grid max-w-2xl auto-rows-fr grid-cols-1 gap-8 sm:mt-20 lg:mx-0 lg:max-w-none lg:grid-cols-3">
+          {PassionProjectsData.map((post) => (
+            <article
+              key={post.id}
+              className="relative flex flex-col justify-end overflow-hidden rounded-2xl bg-gray-900 px-8 pb-8 pt-80 sm:pt-48 lg:pt-80"
+            >
+              <img
+                src={post.imageUrl}
+                alt=""
+                className="absolute inset-0 z-0 h-full w-full object-cover"
+              />
+              <div className="absolute inset-0 z-10 bg-gradient-to-t from-gray-900 via-gray-900/40" />
+              <div className="absolute inset-0 z-10 rounded-2xl ring-1 ring-inset ring-gray-900/10" />
+
+              <div className="flex flex-wrap items-center gap-y-1 overflow-hidden text-sm leading-6 text-gray-300">
+                <time dateTime={post.datetime} className="mr-8">
+                  {post.date}
+                </time>
+                <div className="-ml-4 flex items-center gap-x-4">
+                  <svg
+                    viewBox="0 0 2 2"
+                    className="-ml-0.5 h-0.5 w-0.5 flex-none fill-white/50"
+                  >
+                    <circle cx={1} cy={1} r={1} />
+                  </svg>
+                  <div className="flex gap-x-2.5">
+                    <img
+                      src={post.author.imageUrl}
+                      alt=""
+                      className="h-6 w-6 flex-none rounded-full bg-white/10"
+                    />
+                    {post.author.name}
+                  </div>
+                </div>
+              </div>
+              <h3 className="mt-3 text-lg font-semibold leading-6 text-white">
+                <a href={post.href}>
+                  <span className="absolute inset-0" />
+                  {post.title}
+                </a>
+              </h3>
+            </article>
+          ))}
+        </div>
+      </div>
+    </div>
+  );
+}
